chore: Enabled multiprocessing and garbage collection in grid_search
