+++ main.py
@@ -1,19 +1,13 @@
-from DependencyMiner.miner import DependencyMiner
 from Representation.representation import *
 from Representation.helpers import *
 from Representation.csv_parser import load_truth_table
-from Representation.selection import select_top_k, tournament_selection
-from Representation.sampling import sample_from_TTable
-from Variation_quantale.crossover import VariationQuantale, crossTopOne
-from Variation_quantale.mutation import Mutation
-from FactorGraph_EDA.beta_bp import BetaFactorGraph
+
 from Moses.run_bp_moses import run_bp_moses, _finalize_metapop
 from Moses.run_abp_moses import run_abp_moses
 import random
 import numpy as np
-import math
 from typing import List
-import datetime
+
 
 def run_moses(exemplar: Instance, fitness: FitnessOracle, hyperparams: Hyperparams, 
               knobs: List[Knob], target: List[bool], csv_path: str, 
@@ -68,114 +62,9 @@ def run_moses(exemplar: Instance, fitness: FitnessOracle, hyperparams: Hyperpara
         return final_metapop
 
 
-def grid_search_tuning():
-    print("--- Starting Hyperparameter Grid Search ---")
-    
-    # b_probs = [0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8]
-    # u_probs = [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8]
-    b_probs = [0.5, 0.6, 0.7, 0.8]
-    u_probs = [0.5, 0.6, 0.7, 0.8]
-    
-    
-    random.seed(42)
-    csv_paths = ["example_data/test_parity_4.csv", "example_data/test_parity_5.csv"]
-
-    for csv_path in csv_paths:
-        
-        feature_order = 4 if "parity_4" in csv_path else 5
-        input_data, target = load_truth_table(csv_path, output_col='O')
-        knobs = knobs_from_truth_table(input_data)
-        knobs = [k for k in knobs if k.symbol != 'O']
-        fitness = FitnessOracle(target)
-
-        results = []
-        timestamp = datetime.datetime.now().strftime("%Y%m%d_%H")
-        log_filename = f"grid_search_results_{csv_path[13:-4]}_{timestamp}.txt"
-
-        with open(log_filename, "w") as log_file:
-            header = f"--- Starting Grid Search on {csv_path} at {timestamp} ---\n"
-            print(header.strip())
-            log_file.write(header + "\n")
-            log_file.write(f"{'Bernoulli':<10} | {'Uniform':<10} | {'Score':<10} | {'Top Instance'}\n")
-            log_file.write("-" * 80 + "\n")
-
-
-            for b in b_probs:
-                for u in u_probs:
-                    print(f"\nTesting: Bernoulli={b}, Uniform={u}")
-                    
-                    current_hp = Hyperparams(
-                        mutation_rate=0.3,
-                        crossover_rate=0.5,
-                        num_generations=30,
-                        max_iter=100,
-                        neighborhood_size=50,
-                        fg_type="beta",
-                        bernoulli_prob=b,
-                        uniform_prob=u,
-                        feature_order=feature_order,
-                        initial_population_size=2,
-                        exemplar_selection_size=7,
-                        min_crossover_neighbors=5,
-                        evidence_propagation_steps=30,
-                        max_dist=50,
-                    )
-                    
-                    exemplar = Instance(value=f"(AND)", id=0, score=0.0, knobs=knobs)
-                    exemplar.score = fitness.get_fitness(exemplar)
-                    
-                    metapop = [exemplar]
-                    
-                    final_pop = run_moses(
-                        exemplar=exemplar, 
-                        fitness=fitness, 
-                        hyperparams=current_hp, 
-                        knobs=knobs,
-                        target=target, 
-                        csv_path=csv_path, 
-                        metapop=metapop,
-                        max_iter=current_hp.max_iter,
-                        fg_type=current_hp.fg_type,
-                    )
-                    
-                    # Find best score in this run
-                    if final_pop:
-                        best_inst = max(final_pop, key=lambda x: x.score)
-                        results.append({
-                            'b': b, 
-                            'u': u, 
-                            'score': best_inst.score, 
-                            'instance': best_inst.value
-                        })
-                        print(f"-> Result: Score {best_inst.score:.4f}")
-                        log_line = f"{b:<10.1f} | {u:<10.1f} | {best_inst.score:<10.4f} | {best_inst.value}\n"
-                        log_file.write(log_line)
-                        log_file.flush() # Ensure it's written in case of crash
-                    else:
-                        print("-> Result: No population return")
-                        log_file.write(f"{b:<10.1f} | {u:<10.1f} | {'N/A':<10} | No Population\n")
-                        
-            print("\n--- Tuning Results ---")
-            log_file.write("\n" + "="*80 + "\n")
-            log_file.write("FINAL SUMMARY (Sorted by Score descending)\n")
-            log_file.write("="*80 + "\n")
-
-            results.sort(key=lambda x: x['score'], reverse=True)
-            
-            for r in results:
-                summary_line = f"Score: {r['score']:.4f} | B={r['b']}, U={r['u']} | Inst: {r['instance']}"
-                print(summary_line)
-                log_file.write(summary_line + "\n")
-
-            if results:
-                best = results[0]
-                best_msg = f"\n*** Best Configuration: B={best['b']}, U={best['u']} with Score {best['score']:.4f} ***"
-                print(best_msg)
-                log_file.write(best_msg + "\n")
-
 def main(): 
 
-    np.random.seed(42)
+    # np.random.seed(42)
     # random.seed(42)
     metapop = []
 
@@ -184,7 +73,7 @@ def main():
         mutation_rate=0.3,
         crossover_rate=0.5,
         num_generations=30,
-        max_iter=20,
+        max_iter=100,
         neighborhood_size=20,
         fg_type="beta",
         bernoulli_prob=0.6,
@@ -199,9 +88,7 @@ def main():
     input, target = load_truth_table(csv_path, output_col='O')
     knobs = knobs_from_truth_table(input, exclude='O')
     
-    # s_expr = "(AND (OR (AND A (OR (AND D) (AND (NOT C)))) (AND (NOT D) (OR (AND B) (AND C))) (AND D (NOT A)) (AND (NOT B) A) (AND D C) (AND B C) (AND (NOT B) C)) (OR (AND A) (AND (NOT B) (OR (AND A) (AND (NOT C)))) (AND (NOT D) (NOT C)) (AND C) (AND (NOT B) (NOT A)) (AND A C) (AND (NOT B) (NOT C)) (AND C D)) (OR (AND (NOT A)) (AND (NOT B) (OR (AND A) (AND (NOT C)))) (AND (NOT D) C) (AND (NOT C) D)) (OR (AND C) (AND (NOT D)) (AND (NOT B) (NOT A)) (AND (NOT A) C) (AND B (NOT C)) (AND (NOT A) D) (AND A (NOT D)) (AND C D) (AND (NOT B) (NOT D)) (AND C (NOT D))) (OR (AND (NOT A) (OR (AND B) (AND (NOT C)))) (AND (NOT B) (OR (AND A) (AND (NOT C)))) (AND (NOT D) C) (AND A) (AND (NOT D) B)) (OR (AND (NOT A) (OR (AND B) (AND (NOT C)))) (AND (NOT B) (OR (AND A) (AND (NOT C)))) (AND (NOT D) C) (AND A) (AND (NOT D) B) (AND (NOT A) B) (AND (NOT A) (NOT C)) (AND A D) (AND (NOT A) D) (AND C) (AND A (NOT B)) (AND A (NOT D)) (AND (NOT A) (NOT D)) (AND B (NOT C))) (OR (AND A) (AND (NOT B) (OR (AND (NOT A)) (AND C))) (AND D) (AND (NOT C)) (AND (NOT B) (NOT C)) (AND B D) (AND (NOT B) D)))"  
     exemplar = Instance(value=f"(AND)", id=0, score=0.0, knobs=knobs)
-    # exemplar = Instance(value=s_expr, id=0, score=0.0, knobs=knobs)
     fitness = FitnessOracle(target)
     exemplar.score = fitness.get_fitness(exemplar)
     
@@ -223,4 +110,3 @@ def main():
     
 if __name__ == "__main__":
     main()
-    # grid_search_tuning()
